% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{analysis_dap}
\alias{analysis_dap}
\title{A data analysis plan prepared ahead of data analysis}
\format{
A data frame with 6 rows and 6 variables
}
\usage{
analysis_dap
}
\description{
A data analysis plan prepared ahead of data analysis.
}
\keyword{datasets}
