% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_kobo_analysis.R
\name{auto_kobo_analysis}
\alias{auto_kobo_analysis}
\title{Auto survey from a Kobo tool analysis}
\usage{
auto_kobo_analysis(
  design,
  survey,
  choices,
  group = NULL,
  group_key_sep = " -/- ",
  na_rm = TRUE,
  vartype = "ci",
  level = 0.95,
  choices_sep = "/",
  label_survey = TRUE,
  bind = TRUE
)
}
\arguments{
\item{design}{A srvyr::design object.}

\item{survey}{A survey sheet.}

\item{choices}{A choices sheet.}

\item{group}{A quoted vector of columns to group by. Default to NULL for no group.}

\item{group_key_sep}{A character string to separate grouping column names in a fancy 'group_key' column.}

\item{na_rm}{Should NAs from \code{var} be removed? Default to TRUE.}

\item{vartype}{Report variability as one or more of: standard error ("se", default),
confidence interval ("ci"), variance ("var") or coefficient of variation
("cv").}

\item{level}{(For vartype = "ci" only) A single number or vector of numbers indicating
the confidence level}

\item{choices_sep}{Select multiples choices separator.}

\item{label_survey}{Boolean. Retrieve questions label from the survey sheet? Default to TRUE.}

\item{bind}{Boolean. Output a dataframe or a list?}
}
\value{
A survey-summarized-median data frame
}
\description{
Auto survey from a Kobo tool analysis
}
\section{Automation}{


\link{auto_kobo_analysis} performs a very basic automated analysis using the provided Kobo tool.
}

\seealso{
Other functions for analyzing from Kobo tool: 
\code{\link{kobo_analysis}()},
\code{\link{kobo_analysis_from_dap}()},
\code{\link{kobo_mean}()},
\code{\link{kobo_mean_all}()},
\code{\link{kobo_median}()},
\code{\link{kobo_median_all}()},
\code{\link{kobo_ratio}()},
\code{\link{kobo_select_multiple}()},
\code{\link{kobo_select_multiple_all}()},
\code{\link{kobo_select_one}()},
\code{\link{kobo_select_one_all}()}
}
\concept{functions for analyzing from Kobo tool}
