% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_svy_analysis.R
\name{auto_svy_analysis}
\alias{auto_svy_analysis}
\title{Auto survey analysis}
\usage{
auto_svy_analysis(
  design,
  group = NULL,
  group_key_sep = " ~/~ ",
  na_rm = TRUE,
  vartype = "ci",
  level = 0.95,
  bind = TRUE
)
}
\arguments{
\item{design}{A srvyr::design object.}

\item{group}{A quoted vector of columns to group by. Default to NULL for no group.}

\item{group_key_sep}{A character string to separate grouping column names in a fancy 'group_key' column.}

\item{na_rm}{Should NAs from \code{var} be removed? Default to TRUE.}

\item{vartype}{Report variability as one or more of: standard error ("se", default),
confidence interval ("ci"), variance ("var") or coefficient of variation
("cv").}

\item{level}{A single number or vector of numbers indicating the confidence level}

\item{bind}{Boolean. Output a dataframe or a list?}
}
\value{
A survey-summarized-median data frame
}
\description{
Auto survey analysis
}
\section{Automation}{


\link{auto_svy_analysis} performs a very basic automated analysis. All numeric columns get means and medians. All character columns get a proportion.
}

\seealso{
Other survey analysis functions: 
\code{\link{svy_analysis}()},
\code{\link{svy_interact}()},
\code{\link{svy_mean}()},
\code{\link{svy_median}()},
\code{\link{svy_proportion}()},
\code{\link{svy_quantile}()},
\code{\link{svy_ratio}()}
}
\concept{survey analysis functions}
