% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_indicator.R
\name{check_noncrit_indicator}
\alias{check_noncrit_indicator}
\alias{check_crit_indicator}
\title{Check that indicators contains valid values}
\usage{
check_noncrit_indicator(df, cols_noncrit)

check_crit_indicator(df, cols_crit)
}
\arguments{
\item{df}{A data frame.}

\item{cols_noncrit}{A vector of column names for non-critical indicators (quoted).}

\item{cols_crit}{A vector of column names for critical indicators (quoted).}
}
\value{
A stop statement or \code{TRUE}.
}
\description{
\code{check_noncrit_indicator()} check that the values for  non-critical indicators are either 0, 1 or NA. \code{check_crit_indicator()} check that the values for  non-critical indicators are either 1:5 or NA. Both checks that the columns are integers.
}
