% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{choices}
\alias{choices}
\title{Some choices sheet from a Kobo tool}
\format{
A data frame with 55 rows and 7 variables
}
\usage{
choices
}
\description{
A dataset (in French) containing a basic survey sheet with consensus, admin, enumerator info, chief of hh, and sanitary
}
\keyword{datasets}
