% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{design}
\alias{design}
\title{Some design from the main dataset}
\format{
An object of class \code{tbl_svy} (inherits from \code{survey.design2}, \code{survey.design}) with 3896 rows and 18 columns.
}
\usage{
design
}
\description{
Some design from the main dataset
}
\keyword{datasets}
