% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impactR.analysis-package.R
\docType{package}
\name{impactR.analysis-package}
\alias{impactR.analysis}
\alias{impactR.analysis-package}
\title{impactR.analysis: Analyzing survey data (collected with Kobo)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Providing helpers to support a (quasi-)automated survey data analysis workflow.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://impact-initiatives-hppu.github.io/impactR.analysis/}
}

}
\author{
\strong{Maintainer}: Noblet Guillaume \email{gnoblet@zaclys.net}

}
\keyword{internal}
