% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_analysis.R
\name{kobo_analysis}
\alias{kobo_analysis}
\title{Kobo survey analysis}
\usage{
kobo_analysis(
  design,
  analysis,
  vars,
  survey,
  choices = NULL,
  group = NULL,
  group_key_sep = " -/- ",
  label_survey = TRUE,
  label_choices = TRUE,
  na_rm = TRUE,
  vartype = "ci",
  level = 0.95,
  ratio_key_sep = " -/- ",
  choices_sep = "/"
)
}
\arguments{
\item{design}{A srvyr::design object.}

\item{analysis}{Analysis type. See details for more, well, details.}

\item{vars}{A variable to calculate from.}

\item{survey}{The survey sheet from Kobo (with column "type" split). See \code{split_survey()}.}

\item{choices}{The choices sheet from Kobo.}

\item{group}{A quoted vector of columns to group by. Default to NULL for no group.}

\item{group_key_sep}{A character string to separate grouping column names in a fancy 'group_key' column.}

\item{label_survey}{Boolean. Retrieve var's label from the survey sheet? Default to TRUE.}

\item{label_choices}{Boolean. Retrieve choices label from the choices sheet? Default to TRUE.}

\item{na_rm}{Should NAs from \code{var} be removed? Default to TRUE.}

\item{vartype}{Report variability as one or more of: standard error ("se", default),
confidence interval ("ci"), variance ("var") or coefficient of variation
("cv").}

\item{level}{(For vartype = "ci" only) A single number or vector of numbers indicating
the confidence level}

\item{ratio_key_sep}{A separator for the ratio key.}

\item{choices_sep}{Select multiples choices separator.}
}
\value{
A survey-summarized-median data frame
}
\description{
Kobo survey analysis
}
\section{Specificity per type of analysis}{


The possible analysis type are: mean for \code{\link[=kobo_mean]{kobo_mean()}}, median for \code{\link[=kobo_median]{kobo_median()}}, select_one for \code{\link[=kobo_select_one]{kobo_select_one()}}, select_multiple for \code{\link[=kobo_select_multiple]{kobo_select_multiple()}}, or ratio for \code{\link[=kobo_ratio]{kobo_ratio()}}.
\itemize{
\item If \code{analysis} is "ratio": \code{vars} takes a named vector. Names will be passed to \code{nums} in \code{\link[=svy_ratio]{svy_ratio()}} and values to \code{denoms}.
\item If choices not NULL, the function tries to retrieve labels for select_one, while it is mandatory for select_multiple, and not used for mean, median, and ratio.
}
}

\seealso{
Other functions for analyzing from Kobo tool: 
\code{\link{auto_kobo_analysis}()},
\code{\link{kobo_analysis_from_dap}()},
\code{\link{kobo_mean}()},
\code{\link{kobo_mean_all}()},
\code{\link{kobo_median}()},
\code{\link{kobo_median_all}()},
\code{\link{kobo_ratio}()},
\code{\link{kobo_select_multiple}()},
\code{\link{kobo_select_multiple_all}()},
\code{\link{kobo_select_one}()},
\code{\link{kobo_select_one_all}()}
}
\concept{functions for analyzing from Kobo tool}
