% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_interact.R
\name{kobo_interact}
\alias{kobo_interact}
\title{Get the proportion of the interaction of variables}
\usage{
kobo_interact(
  design,
  vars,
  interact_key_sep = "-/-",
  survey,
  choices = NULL,
  group = NULL,
  group_key_sep = "-/-",
  label_survey = TRUE,
  na_rm = TRUE,
  vartype = "ci",
  level = 0.95
)
}
\arguments{
\item{design}{A srvyr::design object.}

\item{vars}{Variables to calculate proportion from.}

\item{interact_key_sep}{A character string to separate interactions columns in a fancy 'interact_key" column.}

\item{survey}{The survey sheet from Kobo (with column "type" split). See \code{split_survey()}.}

\item{choices}{The choices sheet from Kobo. If not NULL, the function tries to retrieve labels.}

\item{group}{A quoted vector of columns to group by. Default to NULL for no group.}

\item{group_key_sep}{A character string to separate grouping column names in a fancy 'group_key' column.}

\item{label_survey}{Boolean. Retrieve var's label from the survey sheet? Default to TRUE.}

\item{na_rm}{Should NAs from \code{interact} be removed? Default to TRUE.}

\item{vartype}{Report variability as one or more of: standard error ("se", default),
confidence interval ("ci"), variance ("var") or coefficient of variation
("cv").}

\item{level}{(For vartype = "ci" only) A single number or vector of numbers indicating
the confidence level}
}
\description{
A wrapper around \code{\link[=svy_interact]{svy_interact()}} to get the proportion of the interaction of variables, and using the Kobo tool to get labels.
}
