% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_ratio.R
\name{kobo_ratio}
\alias{kobo_ratio}
\title{Get the ratio for numeric variables.}
\usage{
kobo_ratio(
  design,
  nums,
  denoms,
  ratio_key_sep = " -/- ",
  survey,
  group = NULL,
  group_key_sep = " ~/~ ",
  label_survey = TRUE,
  na_rm = TRUE,
  vartype = "ci",
  level = 0.95
)
}
\arguments{
\item{design}{A srvyr::design object.}

\item{nums}{The numerator variables.}

\item{denoms}{The denominator variables.}

\item{ratio_key_sep}{A separator for the ratio key.}

\item{survey}{The survey sheet from Kobo (with column "type" split). See \code{split_survey()}.}

\item{group}{A quoted vector of columns to group by. Default to NULL for no group.}

\item{group_key_sep}{A character string to separate grouping column names in a fancy 'group_key' column.}

\item{label_survey}{Boolean. Retrieve var's label from the survey sheet? Default to TRUE.}

\item{na_rm}{Boolean. Remove any line that as an NA in \code{num} or \code{denom} Default to TRUE.}

\item{vartype}{Report variability as one or more of: standard error ("se", default),
confidence interval ("ci"), variance ("var") or coefficient of variation
("cv").}

\item{level}{A single number or vector of numbers indicating the confidence level}
}
\value{
A character vector of select_one questions.
}
\description{
Get the ratio for numeric variables.
}
\details{
\code{survey} should have a split type column with types of variables such as "select_one", "select_multiple", etc.
}
\seealso{
Other functions for analyzing from Kobo tool: 
\code{\link{auto_kobo_analysis}()},
\code{\link{kobo_analysis}()},
\code{\link{kobo_analysis_from_dap}()},
\code{\link{kobo_mean}()},
\code{\link{kobo_mean_all}()},
\code{\link{kobo_median}()},
\code{\link{kobo_median_all}()},
\code{\link{kobo_select_multiple}()},
\code{\link{kobo_select_multiple_all}()},
\code{\link{kobo_select_one}()},
\code{\link{kobo_select_one_all}()}
}
\concept{functions for analyzing from Kobo tool}
