% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_select_one_all.R
\name{kobo_select_one_all}
\alias{kobo_select_one_all}
\title{Get all select_one proportions}
\usage{
kobo_select_one_all(
  design,
  survey,
  choices = NULL,
  group = NULL,
  group_key_sep = " ~/~ ",
  label_survey = TRUE,
  na_rm = TRUE,
  vartype = "ci",
  level = 0.95
)
}
\arguments{
\item{design}{A srvyr::design object.}

\item{survey}{The survey sheet from Kobo (with column "type" split). See \code{split_survey()}.}

\item{choices}{The choices sheet from Kobo. If not NULL, the function tries to retrieve labels.}

\item{group}{A quoted vector of columns to group by. Default to NULL for no group.}

\item{group_key_sep}{A character string to separate grouping column names in a fancy 'group_key' column.}

\item{label_survey}{Boolean. Retrieve questions label from the survey sheet? Default to TRUE.}

\item{na_rm}{Should NAs from \code{var} be removed? Default to TRUE.}

\item{vartype}{Report variability as one or more of: standard error ("se", default),
confidence interval ("ci"), variance ("var") or coefficient of variation
("cv").}

\item{level}{(For vartype = "ci" only) A single number or vector of numbers indicating
the confidence level}
}
\value{
A character vector of select_one questions.
}
\description{
Get all select_one proportions
}
\details{
\code{survey} should have a split type column with types of variables such as "select_one", "select_multiple", etc.
}
\seealso{
Other functions for analyzing from Kobo tool: 
\code{\link{kobo_mean_all}()},
\code{\link{kobo_mean}()},
\code{\link{kobo_median_all}()},
\code{\link{kobo_median}()},
\code{\link{kobo_ratio}()},
\code{\link{kobo_select_multiple_all}()},
\code{\link{kobo_select_multiple}()},
\code{\link{kobo_select_one}()}
}
\concept{functions for analyzing from Kobo tool}
