% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{main}
\alias{main}
\title{Some MSNA data produced}
\format{
A data frame with 8 rows and 36 variables
}
\usage{
main
}
\description{
The dataset comes from a Kobo tool and pre-processed (weights for instance)
}
\keyword{datasets}
