% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msni}
\alias{msni}
\title{Some composite indicators to compute LSG scores and MSNI score}
\format{
A data frame with 10000 rows and 7 variables
}
\usage{
msni
}
\description{
A dataset with score for non-critical and critical indicators for each sector.
}
\keyword{datasets}
