% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion_select_one.R
\name{proportion_select_one}
\alias{proportion_select_one}
\title{Get the proportion for a select_one}
\usage{
proportion_select_one(
  design,
  col,
  survey,
  choices = NULL,
  group = NULL,
  label_survey = TRUE,
  na_rm = TRUE,
  stat_name = "prop",
  vartype = "ci",
  level = 0.95
)
}
\arguments{
\item{design}{A srvyr::design object.}

\item{col}{A column to calculate proportion from.}

\item{survey}{The survey sheet from Kobo (with column "type" split). See \code{split_survey()}.}

\item{choices}{The choices sheet from Kobo. If not NULL, the function tries to retrieve labels.}

\item{group}{A quoted vector of columns to group by. Default to NULL for no group.}

\item{label_survey}{Boolean. Retrieve questions label from the survey sheet? Default to TRUE.}

\item{na_rm}{Should NAs from \code{col} be removed? Default to TRUE.}

\item{stat_name}{What should the statistic's column be named? Default to "prop".}

\item{vartype}{Report variability as one or more of: standard error ("se", default),
confidence interval ("ci"), variance ("var") or coefficient of variation
("cv").}

\item{level}{(For vartype = "ci" only) A single number or vector of numbers indicating
the confidence level}
}
\value{
A character vector of select_one questions.
}
\description{
Get the proportion for a select_one
}
\details{
\code{survey} should have a split type column with types of variables such as "select_one", "select_multiple", etc.
}
\seealso{
Other functions for analyzing from Kobo tool: 
\code{\link{proportion_all_select_multiple}()},
\code{\link{proportion_all_select_one}()},
\code{\link{proportion_select_multiple}()}
}
\concept{functions for analyzing from Kobo tool}
