% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_indicator.R
\name{score_noncrit_indicator}
\alias{score_noncrit_indicator}
\alias{score_crit_indicator}
\title{Calculate the score for the non-critical indicators component}
\usage{
score_noncrit_indicator(df, ..., new_colname = "score_noncrit")

score_crit_indicator(
  df,
  crit_4plus = NULL,
  crit_4 = NULL,
  crit_3 = NULL,
  new_colname = "score_crit"
)
}
\arguments{
\item{df}{A data frame.}

\item{...}{The variables for the component of non-critical indicators (quoted or not).}

\item{new_colname}{The new column name. It will override it in the dataframe if it already exists.}

\item{crit_4plus}{Vector of critical variables giving up to a severity level of 4+ (quoted).}

\item{crit_4}{Vector of critical variables giving up to a severity level of 4 (quoted).}

\item{crit_3}{Vector of critical variables giving up to a severity level of 3 (quoted).}
}
\value{
A data frame with a new column 'score_noncrit' or 'score_crit'.
}
\description{
\code{score_noncrit_indicator()} calculates the score for the non-critical indicators. \code{score_crit_indicator()} calculates the score for the critical indicators.
}
\details{
A severity of 4+ should be coded as 5 in the dataset.
}
