% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survey}
\alias{survey}
\title{Some survey sheet from a Kobo tool}
\format{
A data frame with 43 rows and 12 variables
}
\usage{
survey
}
\description{
A dataset (in French) containing a basic survey sheet with consensus, admin, enumerator info, chief of hh, and sanitary
}
\keyword{datasets}
