% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_analysis.R
\name{svy_analysis}
\alias{svy_analysis}
\title{Survey analysis}
\usage{
svy_analysis(
  design,
  analysis,
  vars,
  group = NULL,
  group_key_sep = " -/- ",
  na_rm = TRUE,
  vartype = "ci",
  level = 0.95,
  ratio_key_sep = " ~/~ ",
  interact_key_sep = " -/- ",
  quantiles = c(0.25, 0.5, 0.75),
  ak = TRUE,
  ak_overall_sep = " @/@ ",
  ak_main_sep = " -/-",
  ak_var_to_value_sep = " \%/\% ",
  ...
)
}
\arguments{
\item{design}{A srvyr::design object.}

\item{analysis}{Analysis type. See details for more, well, details.}

\item{vars}{A variable to calculate from.}

\item{group}{A quoted vector of columns to group by. Default to NULL for no group.}

\item{group_key_sep}{A character string to separate grouping column names in a fancy 'group_key' column.}

\item{na_rm}{Should NAs from \code{var} be removed? Default to TRUE.}

\item{vartype}{Report variability as one or more of: standard error ("se", default),
confidence interval ("ci"), variance ("var") or coefficient of variation
("cv").}

\item{level}{A single number or vector of numbers indicating the confidence level}

\item{ratio_key_sep}{A separator for the ratio key.}

\item{interact_key_sep}{A character string to separate interactions columns in a fancy 'interact_key" column.}

\item{quantiles}{A vector of quantiles to calculate}

\item{ak}{Boolean. Add the analysis key?}

\item{ak_overall_sep}{The overall separator between items, e.g. between the type of analysis and the variables information.}

\item{ak_main_sep}{The main separator between variables, e.g. between the two grouping columns.}

\item{ak_var_to_value_sep}{The separator between the variable and its value.}

\item{...}{Other parameters to pass to the used \code{srvyr} method.}
}
\value{
A survey-summarized-median data frame
}
\description{
Survey analysis
}
\section{Specificity per type of analysis}{


The possible analysis type are: mean for \code{\link[=svy_mean]{svy_mean()}}, median for \code{\link[=svy_median]{svy_median()}}, interact for \code{\link[=svy_interact]{svy_interact()}}, proportion for \code{\link[=svy_proportion]{svy_proportion()}}, ratio for \code{\link[=svy_ratio]{svy_ratio()}}, or quantile \code{\link[=svy_quantile]{svy_quantile()}}.
\itemize{
\item If \code{analysis} is "ratio": \code{vars} takes a named vector. Names will be passed to \code{nums} in \code{\link[=svy_ratio]{svy_ratio()}} and values to \code{denoms}.
\item If \code{analysis} is "interact": \code{vars} is the equivalent of \code{interact}.
}
}

\seealso{
Other survey analysis functions: 
\code{\link{auto_svy_analysis}()},
\code{\link{svy_interact}()},
\code{\link{svy_mean}()},
\code{\link{svy_median}()},
\code{\link{svy_proportion}()},
\code{\link{svy_quantile}()},
\code{\link{svy_ratio}()}
}
\concept{survey analysis functions}
