% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_mean.R
\name{svy_mean}
\alias{svy_mean}
\title{Survey mean}
\usage{
svy_mean(
  design,
  var,
  group = NULL,
  group_key_sep = "*",
  na_rm = TRUE,
  stat_name = "mean",
  vartype = "ci",
  level = 0.95,
  ...
)
}
\arguments{
\item{design}{A srvyr::design object.}

\item{var}{A variable to calculate the mean from.}

\item{group}{A quoted vector of columns to group by. Default to NULL for no group.}

\item{group_key_sep}{A character string to separate grouping column names in a fancy 'group_key' column.}

\item{na_rm}{Should NAs from \code{var} be removed? Default to TRUE.}

\item{stat_name}{What should the statistic's column be named? Default to "mean".}

\item{vartype}{Report variability as one or more of: standard error ("se", default),
confidence interval ("ci"), variance ("var") or coefficient of variation
("cv").}

\item{level}{(For vartype = "ci" only) A single number or vector of numbers indicating
the confidence level}

\item{...}{Other parameters to pass to \code{srvyr::survey_mean()}.}
}
\value{
A survey-summarized-mean data frame
}
\description{
Survey mean
}
\seealso{
Other survey analysis functions: 
\code{\link{svy_interact}()},
\code{\link{svy_median}()},
\code{\link{svy_prop}()},
\code{\link{svy_quantile}()},
\code{\link{svy_ratio}()}
}
\concept{survey analysis functions}
