% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_median.R
\name{svy_median}
\alias{svy_median}
\title{Survey median}
\usage{
svy_median(
  design,
  vars,
  group = NULL,
  group_key_sep = " ~/~ ",
  na_rm = TRUE,
  vartype = "ci",
  level = 0.95,
  ak = TRUE,
  ak_overall_sep = " @/@ ",
  ak_main_sep = " ~/~ ",
  ak_var_to_value_sep = " \%/\% ",
  ...
)
}
\arguments{
\item{design}{A srvyr::design object.}

\item{vars}{A variable to calculate median from.}

\item{group}{A quoted vector of columns to group by. Default to NULL for no group.}

\item{group_key_sep}{A character string to separate grouping column names in a fancy 'group_key' column.}

\item{na_rm}{Should NAs from \code{var} be removed? Default to TRUE.}

\item{vartype}{NULL to report no variability. Otherwise one or more of: standard error ("se", the default),
confidence interval ("ci"), variance ("var") or coefficient of variation
("cv").}

\item{level}{A single number indicating the confidence level (only one level allowed). Note that this may effect estimated standard errors (see \code{\link[survey]{svyquantile}} details on \code{alpha}, which equals \code{1-level}).}

\item{ak}{Boolean. Add the analysis key?}

\item{ak_overall_sep}{The overall separator between items, e.g. between the type of analysis and the variables information.}

\item{ak_main_sep}{The main separator between variables, e.g. between the two grouping columns.}

\item{ak_var_to_value_sep}{The separator between the variable and its value.}

\item{...}{Other parameters to pass to \code{srvyr::survey_median()}.}
}
\value{
A survey-summarized-median data frame
}
\description{
Survey median
}
\seealso{
Other survey analysis functions: 
\code{\link{auto_svy_analysis}()},
\code{\link{svy_analysis}()},
\code{\link{svy_interact}()},
\code{\link{svy_mean}()},
\code{\link{svy_proportion}()},
\code{\link{svy_quantile}()},
\code{\link{svy_ratio}()}
}
\concept{survey analysis functions}
