% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_quantile.R
\name{svy_quantile}
\alias{svy_quantile}
\title{Survey quantile}
\usage{
svy_quantile(
  design,
  var,
  quantiles = c(0.25, 0.5, 0.75),
  group = NULL,
  group_key_sep = "*",
  na_rm = TRUE,
  vartype = "ci",
  level = 0.95,
  ...
)
}
\arguments{
\item{design}{A srvyr::design object.}

\item{var}{A variable to calculate quantile from.}

\item{quantiles}{A vector of quantiles to calculate}

\item{group}{A quoted vector of columns to group by. Default to NULL for no group.}

\item{group_key_sep}{A character string to separate grouping column names in a fancy 'group_key' column.}

\item{na_rm}{Should NAs from \code{var} be removed? Default to TRUE.}

\item{vartype}{NULL to report no variability. Otherwise one or more of: standard error ("se", the default),
confidence interval ("ci"), variance ("var") or coefficient of variation
("cv").}

\item{level}{A single number indicating the confidence level (only one level allowed). Note that this may effect estimated standard errors (see \code{\link[survey]{svyquantile}} details on \code{alpha}, which equals \code{1-level}).}

\item{...}{Other parameters to pass to \code{srvyr::survey_quantile()}.}
}
\value{
A survey-summarized-mean data frame
}
\description{
Survey quantile
}
\seealso{
Other survey analysis functions: 
\code{\link{svy_interact}()},
\code{\link{svy_mean}()},
\code{\link{svy_median}()},
\code{\link{svy_prop}()},
\code{\link{svy_ratio}()}
}
\concept{survey analysis functions}
