% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_test_chisq.R
\name{svy_test_chisq}
\alias{svy_test_chisq}
\title{Chi-squared survey tests, wrapper around survey::svychisq()}
\usage{
svy_test_chisq(
  design,
  var1,
  var2,
  group = NULL,
  na_rm = TRUE,
  statistic = "F",
  ...
)
}
\arguments{
\item{design}{A srvyr::design object.}

\item{var1}{Var name 1.}

\item{var2}{Var name 2.}

\item{group}{A quoted or unquoted vector of columns to group by. Default to NULL for no group.}

\item{na_rm}{Should NAs from \code{var1} and \code{var2} be removed? Default to TRUE.}

\item{statistic}{See Details below}

\item{...}{Other parameters to pass to \code{survey::svychisq()}.}
}
\value{
A data frame of the test output
}
\description{
Chi-squared survey tests, wrapper around survey::svychisq()
}
\seealso{
Other test functions: 
\code{\link{svy_test_student}()}
}
\concept{test functions}
