% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_test_student.R
\name{svy_test_student}
\alias{svy_test_student}
\title{Student survey tests, wrapper around survey::svyttest()}
\usage{
svy_test_student(
  design,
  var1,
  var2 = NULL,
  group = NULL,
  na_rm = TRUE,
  statistic = "F",
  ...
)
}
\arguments{
\item{design}{A srvyr::design object.}

\item{var1}{Var name 1.}

\item{var2}{Var name 2 for two-sample case. Default to NULL.}

\item{group}{A quoted or unquoted vector of columns to group by. Default to NULL for no group.}

\item{na_rm}{Should NAs from \code{num} and \code{denom} be removed?}

\item{statistic}{See \code{stats::t.test()} for methods.}

\item{...}{for methods }
}
\value{
A survey-summarized-interaction tibble
}
\description{
Student survey tests, wrapper around survey::svyttest()
}
\seealso{
Other test functions: 
\code{\link{svy_test_chisq}()}
}
\concept{test functions}
