% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_analysis_from_dap.R
\name{kobo_analysis_from_dap}
\alias{kobo_analysis_from_dap}
\title{Kobo survey analysis from a data analysis plan}
\usage{
kobo_analysis_from_dap(
  design,
  dap,
  survey,
  choices,
  group = NULL,
  level = 0.95,
  choices_sep = "_"
)
}
\arguments{
\item{design}{A srvyr::design object.}

\item{dap}{A well formatted data analysis plan.#'}

\item{survey}{The survey sheet from Kobo (with column "type" split). See \code{split_survey()}.}

\item{choices}{The choices sheet from Kobo.}

\item{group}{A quoted vector of columns to group by. Default to NULL for no group.}

\item{level}{(For vartype = "ci" only) A single number or vector of numbers indicating
the confidence level}

\item{choices_sep}{Select multiples choices separator.}
}
\value{
A survey-summarized-median data frame
}
\description{
Kobo survey analysis from a data analysis plan
}
\section{Specifics}{

\itemize{
\item It takes the default for most variables, e.g. it tries to label, it calculates confidence intervals, and separators are the default.
\item Choices cannot be NULL.
\item The separator for choices can be changed.
\item Level can be changed.
}
}

\seealso{
Other functions for analyzing from Kobo tool: 
\code{\link{auto_kobo_analysis}()},
\code{\link{kobo_analysis}()},
\code{\link{kobo_mean}()},
\code{\link{kobo_mean_all}()},
\code{\link{kobo_median}()},
\code{\link{kobo_median_all}()},
\code{\link{kobo_ratio}()},
\code{\link{kobo_select_multiple}()},
\code{\link{kobo_select_multiple_all}()},
\code{\link{kobo_select_one}()},
\code{\link{kobo_select_one_all}()}
}
\concept{functions for analyzing from Kobo tool}
