% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_interact.R
\name{svy_interact}
\alias{svy_interact}
\title{Survey interactions}
\usage{
svy_interact(
  design,
  interact,
  interact_key_sep = "*",
  group = NULL,
  group_key_sep = "*",
  unnest_interaction = TRUE,
  na_rm = TRUE,
  vartype = "ci",
  level = 0.95,
  ...
)
}
\arguments{
\item{design}{A srvyr::design object.}

\item{interact}{A quoted vector of variables to calculate interactions from (must be quoted).}

\item{interact_key_sep}{A character string to separate interactions columns in a fancy 'interact_key" column.}

\item{group}{A quoted vector of columns to group by. Default to NULL for no group.}

\item{group_key_sep}{A character string to separate grouping column names in a fancy 'group_key' column.}

\item{unnest_interaction}{Should interaction be unnested? Default to TRUE.}

\item{na_rm}{Should NAs from \code{interact} be removed? Default to TRUE.}

\item{vartype}{Report variability as one or more of: standard error ("se", default),
confidence interval ("ci"), variance ("var") or coefficient of variation
("cv").}

\item{level}{(For vartype = "ci" only) A single number or vector of numbers indicating
the confidence level}

\item{...}{Other parameters to pass to \code{srvyr::survey_mean()}.}
}
\value{
A survey-summarized-mean data frame
}
\description{
Survey interactions
}
\seealso{
Other survey analysis functions: 
\code{\link{svy_mean}()},
\code{\link{svy_median}()},
\code{\link{svy_prop}()},
\code{\link{svy_quantile}()},
\code{\link{svy_ratio}()}
}
\concept{survey analysis functions}
