% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_median.R
\name{svy_median}
\alias{svy_median}
\title{Survey median}
\usage{
svy_median(
  design,
  col,
  group = NULL,
  na_rm = TRUE,
  stat_name = "median",
  vartype = "ci",
  level = 0.95,
  deff = FALSE,
  ...
)
}
\arguments{
\item{design}{A srvyr::design object.}

\item{col}{A column to calculate median from.}

\item{group}{A quoted or unquoted vector of columns to group by. Default to NULL for no group.}

\item{na_rm}{Should NAs from \code{col} be removed? Default to TRUE.}

\item{stat_name}{What should the statistic's column be named? Default to "median".}

\item{vartype}{NULL to report no variability. Otherwise one or more of: standard error ("se", the default),
confidence interval ("ci"), variance ("var") or coefficient of variation
("cv").}

\item{level}{A single number indicating the confidence level (only one level allowed). Note that this may effect estimated standard errors (see \code{\link[survey]{svyquantile}} details on \code{alpha}, which equals \code{1-level}).}

\item{...}{Other parameters to pass to \code{srvyr::survey_median()}.}
}
\value{
A survey-summarized-median data frame
}
\description{
Survey median
}
\seealso{
Other survey analysis functions: 
\code{\link{svy_count_numeric}()},
\code{\link{svy_interact}()},
\code{\link{svy_mean}()},
\code{\link{svy_prop}()},
\code{\link{svy_ratio}()}
}
\concept{survey analysis functions}
