% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_prop.R
\name{svy_prop}
\alias{svy_prop}
\title{Survey proportion}
\usage{
svy_prop(
  design,
  col,
  group = NULL,
  na_rm = TRUE,
  stat_name = "prop",
  vartype = "ci",
  level = 0.95,
  deff = FALSE,
  ...
)
}
\arguments{
\item{design}{A srvyr::design object.}

\item{col}{A column to calculate proportion from.}

\item{group}{A quoted or unquoted vector of columns to group by. Default to NULL for no group.}

\item{na_rm}{Should NAs from \code{col} be removed? Default to TRUE.}

\item{stat_name}{What should the statistic's column be named? Default to "prop".}

\item{vartype}{Report variability as one or more of: standard error ("se", default),
confidence interval ("ci"), variance ("var") or coefficient of variation
("cv").}

\item{level}{(For vartype = "ci" only) A single number or vector of numbers indicating
the confidence level}

\item{deff}{A logical value to indicate whether the design effect should be returned.}

\item{...}{Other parameters to pass to \code{srvyr::survey_prop()}.}
}
\value{
A survey-summarized-proportion data frame
}
\description{
Survey proportion
}
\seealso{
Other survey analysis functions: 
\code{\link{svy_count_numeric}()},
\code{\link{svy_interact}()},
\code{\link{svy_mean}()},
\code{\link{svy_median}()},
\code{\link{svy_ratio}()}
}
\concept{survey analysis functions}
