% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_prop.R
\name{svy_prop}
\alias{svy_prop}
\title{Survey proportion}
\usage{
svy_prop(
  design,
  vars,
  group = NULL,
  group_key_sep = "*",
  na_rm = TRUE,
  vartype = "ci",
  level = 0.95,
  ...
)
}
\arguments{
\item{design}{A srvyr::design object.}

\item{vars}{A quoted vector of variables to calculate proportion from.}

\item{group}{A quoted vector of columns to group by. Default to NULL for no group.}

\item{group_key_sep}{A character string to separate grouping column names in a fancy 'group_key' column.}

\item{na_rm}{Should NAs from \code{var} be removed? Default to TRUE.}

\item{vartype}{Report variability as one or more of: standard error ("se", default),
confidence interval ("ci"), variance ("var") or coefficient of variation
("cv").}

\item{level}{(For vartype = "ci" only) A single number or vector of numbers indicating
the confidence level}

\item{...}{Other parameters to pass to \code{srvyr::survey_prop()}.}
}
\value{
A survey-summarized-proportion data frame
}
\description{
Survey proportion
}
\seealso{
Other survey analysis functions: 
\code{\link{auto_svy_analysis}()},
\code{\link{svy_analysis}()},
\code{\link{svy_interact}()},
\code{\link{svy_mean}()},
\code{\link{svy_median}()},
\code{\link{svy_quantile}()},
\code{\link{svy_ratio}()}
}
\concept{survey analysis functions}
