% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_ratio.R
\name{svy_ratio}
\alias{svy_ratio}
\title{Survey ratio}
\usage{
svy_ratio(
  design,
  num,
  denom,
  group = NULL,
  na_rm = TRUE,
  stat_name = "ratio",
  ...
)
}
\arguments{
\item{design}{A srvyr::design object}

\item{num}{The numerator column}

\item{denom}{The denominator column}

\item{group}{A quoted or unquoted vector of columns to group by. Default to NULL for no group}

\item{na_rm}{Boolean. Remove any line that as an NA in \code{num} or \code{denom} Default to TRUE.}

\item{stat_name}{What should the statistic's column be named? Default to "ratio"}

\item{...}{Parameters to pass to srvyr::survey_mean()}
}
\value{
A survey-summarized-ratio data frame
}
\description{
Survey ratio
}
\seealso{
Other survey analysis functions: 
\code{\link{svy_count_numeric}()},
\code{\link{svy_interact}()},
\code{\link{svy_mean}()},
\code{\link{svy_median}()},
\code{\link{svy_prop}()}
}
\concept{survey analysis functions}
