% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_ratio.R
\name{svy_ratio}
\alias{svy_ratio}
\title{Survey ratio}
\usage{
svy_ratio(
  design,
  nums,
  denoms,
  ratio_key_sep = " / ",
  group = NULL,
  group_key_sep = "*",
  na_rm = TRUE,
  vartype = "ci",
  level = 0.95,
  ...
)
}
\arguments{
\item{design}{A srvyr::design object.}

\item{nums}{The numerator variables.}

\item{denoms}{The denominator variables.}

\item{ratio_key_sep}{A separator for the ratio key.}

\item{group}{A quoted vector of columns to group by. Default to NULL for no group.}

\item{group_key_sep}{A character string to separate grouping column names in a fancy 'group_key' column.}

\item{na_rm}{Boolean. Remove any line that as an NA in \code{num} or \code{denom} Default to TRUE.}

\item{vartype}{Report variability as one or more of: standard error ("se", default),
confidence interval ("ci"), variance ("var") or coefficient of variation
("cv").}

\item{level}{A single number or vector of numbers indicating the confidence level}

\item{...}{Parameters to pass to \code{srvyr::survey_ratio()}.}
}
\value{
A survey-summarized-ratio data frame
}
\description{
Survey ratio
}
\seealso{
Other survey analysis functions: 
\code{\link{auto_svy_analysis}()},
\code{\link{svy_analysis}()},
\code{\link{svy_interact}()},
\code{\link{svy_mean}()},
\code{\link{svy_median}()},
\code{\link{svy_prop}()},
\code{\link{svy_quantile}()}
}
\concept{survey analysis functions}
